refactor(harnesstui): extract reusable transcript reader
