+++ src/loushang/harnesstui/__init__.py
@@ -1,0 +1,1 @@
+"""Product-neutral interaction adapters between Harness and the terminal UI."""
