+++ src/loushang/harnesstui/conversation/__init__.py
@@ -1,0 +1,1 @@
+"""Shared terminal interaction for projected conversations."""
